src/main/java/ru/spbau/lobanov/streamAPI/SecondPartTasks.java
package ru.spbau.lobanov.streamAPI;

import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.Collection;
import java.util.Comparator;
import java.util.List;
import java.util.Map;
import java.util.stream.Collectors;
import java.util.stream.Stream;

public final class SecondPartTasks {

    private SecondPartTasks() {}

    private static Stream<String> getLinesQuiet(Path p) {
        try {
            return Files.lines(p);
        } catch (IOException e) {
            return Stream.empty();
        }
    }

    // Найти строки из переданных файлов, в которых встречается указанная подстрока.
    public static List<String> findQuotes(List<String> paths, CharSequence sequence) {
        return paths.stream().map(Paths::get).filter(Files::isRegularFile)
                .filter(Files::isReadable).flatMap(SecondPartTasks::getLinesQuiet)
                .filter(s -> s.contains(sequence))
                .collect(Collectors.toList());
    }

    // В квадрат с длиной стороны 1 вписана мишень.
    // Стрелок атакует мишень и каждый раз попадает в произвольную точку квадрата.
    // Надо промоделировать этот процесс с помощью класса java.util.Random и посчитать, какова вероятность попасть в мишень.
    public static double piDividedBy4() {
        return Stream.generate(() -> Math.hypot(Math.random(), Math.random()) <= 1)
                .limit(1000000)
                .mapToInt(success -> success? 1 : 0).average().orElse(0);
    }

    // Дано отображение из имени автора в список с содержанием его произведений.
    // Надо вычислить, чья общая длина произведений наибольшая.
    public static String findPrinter(Map<String, List<String>> compositions) {
        return compositions.entrySet().stream().collect(Collectors.maxBy(
                Comparator.comparingInt(e -> e.getValue().stream().mapToInt(String::length).sum())))
                .map(Map.Entry::getKey).orElse("Nobody");
    }

    // Вы крупный поставщик продуктов. Каждая торговая сеть делает вам заказ в виде Map<Товар, Количество>.
    // Необходимо вычислить, какой товар и в каком количестве надо поставить.
    public static Map<String, Integer> calculateGlobalOrder(List<Map<String, Integer>> orders) {
        return orders.stream().map(Map::entrySet).flatMap(Collection::stream)
                .collect(Collectors.groupingBy(Map.Entry::getKey, Collectors.summingInt(Map.Entry::getValue)));
    }
//

}
